# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  parking_slot.AdminUpdateParkingSlotRequest:
    properties:
      status:
        $ref: '#/definitions/parking_slot.SlotStatus'
    required:
    - status
    type: object
  parking_slot.ChangeSlotDeviceRequest:
    properties:
      device_mac:
        type: string
      port_number:
        type: integer
    required:
    - device_mac
    - port_number
    type: object
  parking_slot.CreateParkingSlotRequest:
    properties:
      device_mac:
        type: string
      lot_id:
        type: integer
      name:
        type: string
      port_number:
        type: integer
    required:
    - device_mac
    - lot_id
    - name
    - port_number
    type: object
  parking_slot.SensorUpdateParkingSlotRequest:
    properties:
      is_occupied:
        type: boolean
      mac:
        type: string
      port:
        type: integer
    required:
    - is_occupied
    - mac
    - port
    type: object
  parking_slot.SlotStatus:
    enum:
    - AVAILABLE
    - OCCUPIED
    - MAINTAIN
    type: string
    x-enum-varnames:
    - SlotStatusAvailable
    - SlotStatusOccupied
    - SlotStatusMaintain
  rfid_card.CardType:
    enum:
    - REGISTERED
    - GUEST
    type: string
    x-enum-varnames:
    - CardTypeRegistered
    - CardTypeGuest
  rfid_card.CreateRfidCardRequest:
    properties:
      card_type:
        $ref: '#/definitions/rfid_card.CardType'
      is_active:
        type: boolean
      owner_name:
        type: string
      uid:
        type: string
    required:
    - card_type
    - uid
    type: object
  rfid_card.UpdateRfidCardRequest:
    properties:
      card_type:
        $ref: '#/definitions/rfid_card.CardType'
      is_active:
        type: boolean
      owner_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API cho hệ thống smart parking
  title: Smart Parking Backend API
  version: "1.0"
paths:
...
  /parking-sessions:
    get:
      description: Trả về danh sách tất cả phiên gửi xe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Lấy danh sách phiên gửi xe
      tags:
      - parking_session
  /parking-sessions/{id}:
    delete:
      description: Đóng phiên gửi xe theo ID, dùng cho testing
      parameters:
      - description: ID phiên gửi xe
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đóng phiên gửi xe theo ID (test)
      tags:
      - parking_session
    get:
      description: Lấy thông tin phiên gửi xe theo ID
      parameters:
      - description: ID phiên gửi xe
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Lấy chi tiết phiên gửi xe
      tags:
      - parking_session
  /parking-sessions/card/{uid}:
    delete:
      description: Đóng phiên gửi xe đang active theo UID thẻ, dùng cho testing
      parameters:
      - description: UID thẻ
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Đóng phiên gửi xe theo UID thẻ (test)
      tags:
      - parking_session
  /parking-sessions/purge/{id}:
    delete:
      description: Xóa vĩnh viễn phiên gửi xe khỏi DB
      parameters:
      - description: ID phiên gửi xe
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Xóa vĩnh viễn phiên gửi xe
      tags:
      - parking_session
  /parking-slots:
    post:
      consumes:
      - application/json
      description: Tạo mới một vị trí đỗ xe trong hệ thống
      parameters:
      - description: Thông tin vị trí đỗ xe
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/parking_slot.CreateParkingSlotRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tạo vị trí đỗ xe
      tags:
      - parking_slot
  /parking-slots/{id}:
    get:
      description: Lấy thông tin chi tiết của một vị trí đỗ xe theo ID
      parameters:
      - description: ID vị trí đỗ xe
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lấy chi tiết vị trí đỗ xe
      tags:
      - parking_slot
  /parking-slots/{id}/device:
    patch:
      consumes:
      - application/json
      description: Cập nhật hoặc thay đổi thiết bị IoT gắn với vị trí đỗ xe theo ID
      parameters:
      - description: ID vị trí đỗ xe
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin thiết bị mới
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/parking_slot.ChangeSlotDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Đổi thiết bị cho vị trí đỗ
      tags:
      - parking_slot
  /parking-slots/admin/{id}:
    patch:
      consumes:
      - application/json
      description: Cập nhật trạng thái vị trí đỗ xe theo ID bởi quản trị viên hoặc
        quản lý
      parameters:
      - description: ID vị trí đỗ xe
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin cập nhật trạng thái
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/parking_slot.AdminUpdateParkingSlotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Quản trị viên cập nhật trạng thái vị trí đỗ
      tags:
      - parking_slot
  /parking-slots/sensor:
    post:
      consumes:
      - application/json
      description: Cập nhật trạng thái vị trí đỗ xe từ cảm biến hoặc thiết bị IoT
      parameters:
      - description: Dữ liệu trạng thái từ cảm biến
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/parking_slot.SensorUpdateParkingSlotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Thiết bị cảm biến cập nhật trạng thái vị trí đỗ
      tags:
      - parking_slot
  /rfid-cards:
    post:
      consumes:
      - application/json
      description: Tạo mới thẻ RFID trong hệ thống
      parameters:
      - description: Thông tin thẻ RFID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rfid_card.CreateRfidCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tạo thẻ RFID
      tags:
      - rfid_card
  /rfid-cards/{id}:
    patch:
      consumes:
      - application/json
      description: Cập nhật thông tin thẻ RFID theo ID
      parameters:
      - description: ID thẻ RFID
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rfid_card.UpdateRfidCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật thẻ RFID
      tags:
      - rfid_card
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"